+++ examples/asa/scripts/2-gold-asc.js
@@ -10,19 +10,21 @@ async function run (runtimeEnv, deployer) {
   await executeTransaction(deployer, mkParam(masterAccount, goldOwner.addr, 200e6, { note: 'funding account' }));
 
   // save Smart Signature by name & fund the account
-  const ascInfoContract = await deployer.mkContractLsig('2-gold-contract-asc.teal', {}, 'Gold_C_Lsig');
+  const ascInfoContract = await deployer.mkContractLsig('Gold_C_Lsig', '2-gold-contract-asc.teal', {});
   console.log(ascInfoContract);
-  await deployer.fundLsigByName('Gold_C_Lsig',
+  await deployer.fundLsig('Gold_C_Lsig',
     { funder: goldOwner, fundingMicroAlgo: 1e6 }, {}); // funding with 1 Algo
 
-  const ascInfoAlgoDelegated = await deployer.mkDelegatedLsig('3-gold-delegated-asc.teal', goldOwner);
-  const ascInfoGoldDelegated = await deployer.mkDelegatedLsig('4-gold-asa.teal', goldOwner);
+  const ascInfoAlgoDelegated =
+    await deployer.mkDelegatedLsig('Gold_D_Lsig', '3-gold-delegated-asc.teal', goldOwner);
+  const ascInfoGoldDelegated =
+    await deployer.mkDelegatedLsig('Gold_d_asa_lsig', '4-gold-asa.teal', goldOwner);
 
   console.log(ascInfoAlgoDelegated);
   console.log(ascInfoGoldDelegated);
 
   /* Contract opt-in for ASA gold + fund contract with ASA gold */
-  const lsig = await deployer.loadLogic('2-gold-contract-asc.teal');
+  const lsig = await deployer.getLsig('Gold_C_Lsig');
   const goldAsset = deployer.asa.get('gold');
   const goldAssetID = goldAsset.assetIndex;
   await deployer.optInLsigToASA(goldAssetID, lsig, { totalFee: 1000 });