+++ examples/permissioned-token-freezing/scripts/1-assetConfig.js
@@ -7,7 +7,7 @@ async function run (runtimeEnv, deployer) {
   const creator = deployer.accountsByName.get('alice');
 
   // NOTE: make sure to deploy 0-createAppAsset.js first
-  const appInfo = deployer.getApp('poi-approval.teal', 'poi-clear.teal');
+  const appInfo = deployer.getApp('PermissionedTokenApp');
   const assetInfo = deployer.asa.get('gold');
 
   /** * Compile and fund escrow***/
@@ -15,11 +15,12 @@ async function run (runtimeEnv, deployer) {
     ASSET_ID: assetInfo.assetIndex,
     APP_ID: appInfo.appID
   };
+  await deployer.mkContractLsig('clawbackEscrow', 'clawback-escrow.py', escrowParams);
 
-  await deployer.fundLsig('clawback-escrow.py',
-    { funder: creator, fundingMicroAlgo: 1e6 }, {}, escrowParams); // sending 1 Algo
+  await deployer.fundLsig('clawbackEscrow',
+    { funder: creator, fundingMicroAlgo: 1e6 }, {}); // sending 1 Algo
 
-  const escrowLsig = await deployer.loadLogic('clawback-escrow.py', escrowParams);
+  const escrowLsig = deployer.getLsig('clawbackEscrow');
   const escrowAddress = escrowLsig.address();
 
   /** Update clawback address to escrow **/