+++ packages/algob/test/fixture-projects/scripts-dir/scripts/other-scripts/deploy-asc.js
@@ -2,9 +2,9 @@
 const fs = require('fs')
 
 async function run (runtimeEnv, deployer) {
-  fs.appendFileSync('output.txt', 'fundLsig script\n')
-  await deployer.fundLsig('metadata key', {}, 'metadata value')
-  fs.appendFileSync('output.txt', 'fundLsig script after\n')
+  fs.appendFileSync('output.txt', 'fundLsigByFile script\n')
+  await deployer.fundLsigByFile('metadata key', {}, 'metadata value')
+  fs.appendFileSync('output.txt', 'fundLsigByFile script after\n')
 }
 
 module.exports = { default: run }