+++ packages/algob/test/lib/tx.ts
@@ -251,7 +251,7 @@ describe("Delete ASA and SSC", () => {
 
     await executeTransaction(deployer, execParams);
 
-    const res = deployer.getApp("approval.teal", "clear.teal");
+    const res = deployer.getAppByFile("approval.teal", "clear.teal");
     assert.isDefined(res);
     if (res) assert.equal(res.deleted, true);
   });
@@ -640,7 +640,10 @@ describe("Deploy, Delete transactions test in run mode", () => {
     await executeTransaction(deployer, execParams);
 
     // should not be stored in checkpoint if in run mode
-    assert.isUndefined(deployer.getApp("approval.teal", "clear.teal"));
+    expectBuilderError(
+      () => deployer.getAppByFile("approval.teal", "clear.teal"),
+      ERRORS.GENERAL.APP_NOT_FOUND_IN_CP
+    );
   });
 
   it("should deploy application in deploy mode and save info by name", async () => {
@@ -661,10 +664,13 @@ describe("Deploy, Delete transactions test in run mode", () => {
     await executeTransaction(deployer, execParams);
 
     // able to retrieve info by "appName"
-    assert.isDefined(deployer.getAppByName("dao-app"));
+    assert.isDefined(deployer.getApp("dao-app"));
 
     // do note that traditional way doesn't work if appName is passed
-    assert.isUndefined(deployer.getApp("approval.teal", "clear.teal"));
+    expectBuilderError(
+      () => deployer.getAppByFile("approval.teal", "clear.teal"),
+      ERRORS.GENERAL.APP_NOT_FOUND_IN_CP
+    );
   });
 
   it("should delete application in run mode", async () => {
@@ -694,7 +700,7 @@ describe("Deploy, Delete transactions test in run mode", () => {
 
     await executeTransaction(deployer, execParams);
 
-    const res = deployer.getApp("approval.teal", "clear.teal");
+    const res = deployer.getAppByFile("approval.teal", "clear.teal");
     assert.isDefined(res);
     assert.equal(res?.deleted, false);
   });
@@ -734,7 +740,10 @@ describe("Update transaction test in run mode", () => {
     const appInfo = await executeTransaction(deployer, execParams);
 
     // should not be stored in checkpoint if in run mode
-    assert.isUndefined(deployer.getApp("approval.teal", "clear.teal"));
+    expectBuilderError(
+      () => deployer.getAppByFile("approval.teal", "clear.teal"),
+      ERRORS.GENERAL.APP_NOT_FOUND_IN_CP
+    );
 
     execParams = {
       type: wtypes.TransactionType.UpdateApp,
@@ -748,7 +757,10 @@ describe("Update transaction test in run mode", () => {
 
     await executeTransaction(deployer, execParams);
     // should not be stored in checkpoint if in run mode
-    assert.isUndefined(deployer.getApp("approval.teal", "clear.teal"));
+    expectBuilderError(
+      () => deployer.getAppByFile("approval.teal", "clear.teal"),
+      ERRORS.GENERAL.APP_NOT_FOUND_IN_CP
+    );
   });
 
   it("deploy in deploy mode, update in run mode", async () => {
@@ -766,7 +778,7 @@ describe("Update transaction test in run mode", () => {
       payFlags: {}
     };
     await executeTransaction(deployer, execParams);
-    const appInfo = deployer.getApp("approval.teal", "clear.teal");
+    const appInfo = deployer.getAppByFile("approval.teal", "clear.teal");
     assert.isDefined(appInfo);
 
     deployer = new DeployerRunMode(deployerCfg);
@@ -782,7 +794,7 @@ describe("Update transaction test in run mode", () => {
       };
 
       await executeTransaction(deployer, execParams);
-      assert.deepEqual(appInfo, deployer.getApp("approval.teal", "clear.teal"));
+      assert.deepEqual(appInfo, deployer.getAppByFile("approval.teal", "clear.teal"));
     }
   });
 
@@ -800,7 +812,10 @@ describe("Update transaction test in run mode", () => {
       payFlags: {}
     };
     const appInfo = await executeTransaction(deployer, execParams);
-    assert.isUndefined(deployer.getApp("approval.teal", "clear.teal"));
+    expectBuilderError(
+      () => deployer.getAppByFile("approval.teal", "clear.teal"),
+      ERRORS.GENERAL.APP_NOT_FOUND_IN_CP
+    );
 
     deployer = new DeployerDeployMode(deployerCfg);
     execParams = {
@@ -815,7 +830,7 @@ describe("Update transaction test in run mode", () => {
 
     await executeTransaction(deployer, execParams);
     // checkpoint is stored for the update
-    assert.isDefined(deployer.getApp("approval.teal", "clear.teal"));
+    assert.isDefined(deployer.getAppByFile("approval.teal", "clear.teal"));
   });
 });